Fix numeric query parameters (#6)
Datasette's JSON api doesn't provide a way to send query parameters that
are treated as integers/floats rather than strings. To work around this,
we do the dumb/bad thing and just do string substitutions. This is
_less_ bad than it may first seem:

- The subbed in parameter names are unlikely to occur in a real SQL
  query except as parameters. Not 100% reassuring, but for a quick side
  project I'm fine with this.
- We're only subbing in ints and floats, _not_ strings. This means that
  while an error here may generate invalid SQL or wrong results, it
  shouldn't be possible (AFAICT) to do anything super malicious.

If this wasn't a fun side project, I'd nix this solution entirely.
Hopefully we can rip this code out for something more robust in the
future.